refactor(web): split locationService.ts (529 lines) into normalizer and detection modules
Extract into two focused, single-responsibility files:

- locationNormalizer.ts (145 lines): normalizeToAppLocation, buildAppLocation,
  reverseGeocode, normalizeLocationName, formatLocation — pure data
  transforms with no browser API dependencies, safe for SSR
- locationDetection.ts (217 lines): LocationDetect* types, state machine,
  detectPreciseLocationGenerator, getCurrentLocationResult, IP fallback —
  browser GPS domain with retry logic and exponential backoff

locationService.ts reduced to a 37-line re-export barrel preserving
all 10 import sites. Zero changes required at any call site.

File size: 529 → 37 lines (93% reduction at entry point)
