+++ apps/web/src/lib/location/locationService.ts
@@ -1,524 +1,32 @@
-import type { AppLocation, AppLocationSource, GeoJSONPoint } from "@/types/location";
-import { DEFAULT_APP_LOCATION } from "@/types/location";
-import {
-    reverseGeocode as reverseGeocodeApi,
-} from "@/lib/api/user/locations";
-import { detectLocationByIP } from "@/lib/api/ipGeolocation";
-import { createPoint, toCanonicalGeoPoint } from "@esparex/shared";
-import {
-    getDisplayLocationLabel,
-    getHeaderLocationText,
-    getSearchLocationLabel,
-    isGenericDetectedLocation,
-    normalizeLocationText,
-    sanitizeLocationLabel,
-    LABEL_CURRENT_LOCATION,
-    LABEL_CURRENT_LOCATION_CAPTURED,
-} from "@/lib/location/locationLabels";
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-    typeof value === "object" && value !== undefined && !Array.isArray(value);
-
-const normalizeSource = (value: unknown): AppLocationSource => {
-    if (value === "auto" || value === "ip" || value === "manual" || value === "default") {
-        return value;
-    }
-    return "manual";
-};
-
-function buildAppLocation(params: {
-    formattedAddress?: string;
-    city?: string;
-    state?: string;
-    country?: string;
-    pincode?: string;
-    coordinates?: GeoJSONPoint;
-    source?: AppLocationSource;
-    locationId?: string;
-    name?: string;
-    level?: AppLocation["level"];
-    isSnapped?: boolean;
-}): AppLocation {
-    const source = params.source ?? "manual";
-    const city = params.city || DEFAULT_APP_LOCATION.city;
-    const state = params.state || DEFAULT_APP_LOCATION.state;
-    const country = params.country || DEFAULT_APP_LOCATION.country;
-    const formattedAddress =
-        params.formattedAddress ||
-        params.name ||
-        (city && state ? `${city}, ${state}` : city) ||
-        (state ? `${state}, ${country}` : (country !== "India" ? country : ""));
-
-    const locationId = params.locationId;
-    const now = Date.now();
-
-    return {
-        formattedAddress,
-        city,
-        state,
-        country,
-        pincode: params.pincode,
-        source,
-        locationId,
-        level: params.level,
-        id: locationId,
-        display: formattedAddress,
-        coordinates: params.coordinates,
-        detectedAt: now,
-        isAuto: source === "auto",
-        isSnapped: params.isSnapped,
-    };
-}
-
-export function normalizeToAppLocation(
-    rawLocation: unknown,
-    sourceOverride?: AppLocationSource
-): AppLocation | null {
-    if (!rawLocation) return null;
-
-    if (typeof rawLocation === "string") {
-        const value = rawLocation.trim();
-        if (!value) return null;
-        const parts = value.split(",");
-        const city = parts[0]?.trim() || value;
-        const state = parts[1]?.trim() || city;
-        return buildAppLocation({
-            formattedAddress: value,
-            city,
-            state,
-            source: sourceOverride ?? "manual",
-            name: city,
-        });
-    }
-
-    if (!isRecord(rawLocation)) return null;
-
-    const coordinates =
-        toCanonicalGeoPoint(rawLocation.coordinates) ||
-        toCanonicalGeoPoint(rawLocation.location) ||
-        toCanonicalGeoPoint(rawLocation);
-
-    const city =
-        (typeof rawLocation.city === "string" && rawLocation.city) ||
-        (typeof rawLocation.name === "string" && rawLocation.name) ||
-        "";
-
-    const state =
-        (typeof rawLocation.state === "string" && rawLocation.state) || city;
-    const country =
-        (typeof rawLocation.country === "string" && rawLocation.country) ||
-        DEFAULT_APP_LOCATION.country;
-
-    const formattedAddress =
-        (typeof rawLocation.formattedAddress === "string" &&
-            rawLocation.formattedAddress) ||
-        (typeof rawLocation.display === "string" && rawLocation.display) ||
-        (typeof rawLocation.address === "string" && rawLocation.address) ||
-        (typeof rawLocation.name === "string" && rawLocation.name) ||
-        city;
-
-    const locationId =
-        (typeof rawLocation.locationId === "string" && rawLocation.locationId) ||
-        (typeof rawLocation.id === "string" && rawLocation.id) ||
-        undefined;
-
-    const source = sourceOverride ?? normalizeSource(rawLocation.source);
-    const pincode =
-        typeof rawLocation.pincode === "string" ? rawLocation.pincode : undefined;
-    const level =
-        rawLocation.level === "country" ||
-            rawLocation.level === "state" ||
-            rawLocation.level === "district" ||
-            rawLocation.level === "city" ||
-            rawLocation.level === "area" ||
-            rawLocation.level === "village"
-            ? rawLocation.level
-            : undefined;
-
-    const isSnapped = typeof rawLocation.isSnapped === "boolean" ? rawLocation.isSnapped : undefined;
-
-    return buildAppLocation({
-        formattedAddress,
-        city,
-        state,
-        country,
-        pincode,
-        coordinates,
-        source,
-        locationId,
-        level,
-        isSnapped,
-        name:
-            (typeof rawLocation.name === "string" && rawLocation.name) || city,
-    });
-}
-
-export async function reverseGeocode(
-    latitude: number,
-    longitude: number
-): Promise<AppLocation | null> {
-    const location = await reverseGeocodeApi(latitude, longitude);
-    if (!location) return null;
-    return normalizeToAppLocation(location, "auto");
-}
-
-type CurrentLocationOptions = {
-    allowApproximateFallback?: boolean;
-    timeoutMs?: number;
-    maximumAgeMs?: number;
-    enableHighAccuracy?: boolean;
-};
-
-export type LocationDetectFailureReason =
-    | "permission_denied"
-    | "position_unavailable"
-    | "timeout"
-    | "unsupported"
-    | "insecure_context"
-    | "prompt_skipped"
-    | "unknown";
-
-export type LocationDetectFailure = {
-    reason: LocationDetectFailureReason;
-    message: string;
-};
-
-export type LocationDetectResult = {
-    location: AppLocation | null;
-    source: "auto" | "ip" | "none";
-    failure?: LocationDetectFailure;
-};
-
-export type LocationDetectionState =
-    | "idle"
-    | "checking_permission"
-    | "requesting_gps"
-    | "gps_acquired"
-    | "reverse_geocoding"
-    | "location_resolved"
-    | "permission_denied"
-    | "gps_timeout"
-    | "reverse_geocode_failed"
-    | "network_error";
-
-export const LOCATION_STATE_MESSAGES: Record<LocationDetectionState, string> = {
-    idle: "Detect My Location",
-    checking_permission: "Checking location permission...",
-    requesting_gps: "Requesting GPS access...",
-    gps_acquired: "GPS location acquired...",
-    reverse_geocoding: "Finding your nearest city...",
-    location_resolved: "Location detected",
-    permission_denied: "Location permission denied",
-    gps_timeout: "GPS request timed out",
-    reverse_geocode_failed: "Unable to determine your location",
-    network_error: "Unable to connect to location service",
-};
-
-const mapGeolocationError = (error: unknown): LocationDetectFailure => {
-    const code = isRecord(error) && typeof error.code === "number" ? error.code : null;
-    if (code === 1) {
-        return {
-            reason: "permission_denied",
-            message:
-                "Location permission denied. Allow location access in your browser settings and try again.",
-        };
-    }
-    if (code === 2) {
-        return {
-            reason: "position_unavailable",
-            message: "Location unavailable. Check GPS or network and try again.",
-        };
-    }
-    if (code === 3) {
-        return {
-            reason: "timeout",
-            message: "Location request timed out. Please try again.",
-        };
-    }
-    return {
-        reason: "unknown",
-        message: "Unable to detect location right now. Please try again.",
-    };
-};
-
-const isSecureLocationContext = (): boolean => {
-    if (typeof window === "undefined") return false;
-    if (window.isSecureContext) return true;
-    return /^(localhost|127\.0\.0\.1|\[::1\])$/i.test(window.location.hostname);
-};
-
 /**
- * Auto-detect location using the canonical reverse-geocode path only.
- * The location ingest endpoint is admin-only and intentionally not used by the user app.
+ * locationService.ts — Re-export barrel (backward-compatible facade).
+ *
+ * Previously 529 lines. Split into two focused modules:
+ *   - locationNormalizer.ts  — pure data transforms (SSR-safe)
+ *   - locationDetection.ts   — browser GPS/IP detection (browser-only)
+ *
+ * All existing imports from "@/lib/location/locationService" continue to work.
  */
-const autoDetectLocation = async (
-    latitude: number,
-    longitude: number
-): Promise<AppLocation | null> => {
-    const existing = await reverseGeocodeApi(latitude, longitude);
-    if (existing) {
-        return normalizeToAppLocation(existing, "auto");
-    }
 
-    return null;
-};
+export {
+    normalizeToAppLocation,
+    reverseGeocode,
+    normalizeLocationName,
+    formatLocation,
+} from "./locationNormalizer";
 
-const buildFailureResult = (
-    failure: LocationDetectFailure
-): LocationDetectResult => ({
-    location: null,
-    source: "none",
-    failure,
-});
+export type {
+    LocationDetectFailureReason,
+    LocationDetectFailure,
+    LocationDetectResult,
+    LocationDetectionState,
+} from "./locationDetection";
 
-const detectApproximateLocationByIP = async (): Promise<LocationDetectResult | null> => {
-    const detected = await detectLocationByIP();
-    if (!detected?.city || !detected.coordinates) {
-        return null;
-    }
-
-    const formattedAddress = [detected.city, detected.state, detected.country]
-        .filter((value) => typeof value === "string" && value.trim().length > 0)
-        .join(", ");
-
-    return {
-        location: buildAppLocation({
-            formattedAddress: formattedAddress || "Approximate current location",
-            city: detected.city,
-            state: detected.state,
-            country: detected.country,
-            coordinates: detected.coordinates,
-            source: "ip",
-            name: detected.city,
-        }),
-        source: "ip",
-    };
-};
-
-export async function* detectPreciseLocationGenerator(
-    options: CurrentLocationOptions = {}
-): AsyncGenerator<LocationDetectionState, LocationDetectResult, void> {
-    const {
-        allowApproximateFallback = true,
-        timeoutMs = 20000,
-        maximumAgeMs = 0,
-        enableHighAccuracy = true,
-    } = options;
-
-    yield "checking_permission";
-
-    if (!isSecureLocationContext()) {
-        yield "permission_denied";
-        const approximate = allowApproximateFallback
-            ? await detectApproximateLocationByIP()
-            : null;
-        if (approximate) {
-            return approximate;
-        }
-        return buildFailureResult({
-            reason: "insecure_context",
-            message:
-                "Location permission is blocked on insecure pages. Use HTTPS or localhost.",
-        });
-    }
-
-    if (!navigator.geolocation) {
-        yield "permission_denied";
-        const approximate = allowApproximateFallback
-            ? await detectApproximateLocationByIP()
-            : null;
-        if (approximate) {
-            return approximate;
-        }
-
-        return buildFailureResult({
-            reason: "unsupported",
-            message: "Geolocation is not supported by this browser.",
-        });
-    }
-
-    // We intentionally DO NOT await navigator.permissions.query() here.
-    // Awaiting promises and crossing React state boundaries can cause the browser 
-    // to lose the "user gesture" context. If the gesture is lost, modern browsers 
-    // (like Chrome/Safari) will silently suppress the geolocation popup, causing 
-    // an indefinite hang. getCurrentPosition natively handles "denied" anyway by 
-    // immediately returning error.code === 1.
-
-    // PR-3 Hardening: 3 attempts with exponential backoff for transient "Unknown" errors
-    const MAX_ATTEMPTS = 3;
-    let lastError: unknown = undefined;
-    let lastFailureReason: LocationDetectFailureReason = "unknown";
-
-    for (let attempt = 1; attempt <= MAX_ATTEMPTS; attempt++) {
-        try {
-            yield "requesting_gps";
-            const coords = await new Promise<GeolocationCoordinates>(
-                (resolve, reject) => {
-                    navigator.geolocation.getCurrentPosition(
-                        (position) => resolve(position.coords),
-                        reject,
-                        {
-                            enableHighAccuracy,
-                            timeout: timeoutMs,
-                            maximumAge: attempt === 1 ? maximumAgeMs : 0, // Force fresh on retry
-                        }
-                    );
-                }
-            );
-
-            yield "gps_acquired";
-            yield "reverse_geocoding";
-            const resolved = await autoDetectLocation(
-                coords.latitude,
-                coords.longitude
-            );
-            
-            if (resolved) {
-                yield "location_resolved";
-                return {
-                    location: resolved,
-                    source: "auto",
-                };
-            }
-
-            yield "reverse_geocode_failed";
-            return {
-                location: buildAppLocation({
-                    formattedAddress: LABEL_CURRENT_LOCATION_CAPTURED,
-                    city: LABEL_CURRENT_LOCATION,
-                    state: "",
-                    country: "Unknown",
-                    coordinates: createPoint(coords.longitude, coords.latitude),
-                    source: "auto",
-                    name: LABEL_CURRENT_LOCATION,
-                }),
-                source: "auto",
-            };
-        } catch (error) {
-            lastError = error;
-            const failure = mapGeolocationError(error);
-            lastFailureReason = failure.reason;
-
-            // Only retry on transient hardware/soft errors (code 2 = Unavailable, code 3 = Timeout)
-            // Permission Denied (code 1) should fail immediately.
-            const isTransient = failure.reason === "position_unavailable" || failure.reason === "timeout";
-            
-            if (!isTransient || attempt === MAX_ATTEMPTS) {
-                break;
-            }
-
-            // Exponential backoff: 800ms, 1600ms
-            const delay = 400 * Math.pow(2, attempt);
-            await new Promise(r => setTimeout(r, delay));
-        }
-    }
-
-    // High-accuracy (GPS) failed with position_unavailable (kCLErrorLocationUnknown on macOS).
-    // This is common on desktops/laptops that cannot get a satellite fix indoors.
-    // Try low-accuracy mode which uses WiFi/network triangulation via the browser API —
-    // this succeeds in most indoor environments where GPS does not.
-    if (lastFailureReason === "position_unavailable" && enableHighAccuracy) {
-        try {
-            yield "requesting_gps";
-            const coords = await new Promise<GeolocationCoordinates>((resolve, reject) => {
-                navigator.geolocation.getCurrentPosition(
-                    (position) => resolve(position.coords),
-                    reject,
-                    {
-                        enableHighAccuracy: false, // Use WiFi/network triangulation
-                        timeout: 10000,
-                        maximumAge: 60000,
-                    }
-                );
-            });
-
-            yield "gps_acquired";
-            yield "reverse_geocoding";
-            const resolved = await autoDetectLocation(coords.latitude, coords.longitude);
-            
-            if (resolved) {
-                yield "location_resolved";
-                return { location: resolved, source: "auto" };
-            }
-
-            yield "reverse_geocode_failed";
-            return {
-                location: buildAppLocation({
-                    formattedAddress: LABEL_CURRENT_LOCATION_CAPTURED,
-                    city: LABEL_CURRENT_LOCATION,
-                    state: "",
-                    country: "Unknown",
-                    coordinates: createPoint(coords.longitude, coords.latitude),
-                    source: "auto",
-                    name: LABEL_CURRENT_LOCATION,
-                }),
-                source: "auto",
-            };
-        } catch {
-            // Low-accuracy also failed — fall through to IP detection
-        }
-    }
-
-    const failure = mapGeolocationError(lastError);
-    if (failure.reason === "permission_denied") {
-        yield "permission_denied";
-    } else if (failure.reason === "timeout") {
-        yield "gps_timeout";
-    } else {
-        yield "network_error";
-    }
-    
-    if (allowApproximateFallback && failure.reason !== "permission_denied") {
-        const approximate = await detectApproximateLocationByIP();
-        if (approximate) {
-            yield "location_resolved";
-            return approximate;
-        }
-    }
-
-    return buildFailureResult(failure);
-}
-
-export async function getCurrentLocationResult(
-    options: CurrentLocationOptions = {}
-): Promise<LocationDetectResult> {
-    const generator = detectPreciseLocationGenerator(options);
-    let finalResult: LocationDetectResult | undefined;
-    while (true) {
-        const { value, done } = await generator.next();
-        if (done) {
-            finalResult = value as LocationDetectResult;
-            break;
-        }
-    }
-    return finalResult;
-}
-
-export function normalizeLocationName(name: string | undefined | null): string {
-    // Display-only formatter.
-    // Must not be used in backend/search query construction to avoid
-    // altering canonical lookup behavior for diacritic-sensitive names.
-    return normalizeLocationText(name);
-}
-
-type LocationLike = {
-    display?: string;
-    city?: string;
-    name?: string;
-} | string | null | undefined;
-
-export function formatLocation(location: LocationLike): string {
-    if (!location) return "";
-    if (typeof location === "string") return sanitizeLocationLabel(location) || "";
-    
-    // Prioritize City for brief "brief" indicators (e.g. Ad Cards).
-    // The "display" field often contains full addresses which causes grid inconsistency.
-    if (location.city) return sanitizeLocationLabel(location.city) || "";
-    if (location.display) return sanitizeLocationLabel(location.display) || "";
-    if (location.name) return sanitizeLocationLabel(location.name) || "";
-
-    return "";
-}
+export {
+    LOCATION_STATE_MESSAGES,
+    detectPreciseLocationGenerator,
+    getCurrentLocationResult,
+} from "./locationDetection";
 
 export {
     getDisplayLocationLabel,
@@ -526,4 +34,4 @@ export {
     getSearchLocationLabel,
     isGenericDetectedLocation,
     sanitizeLocationLabel,
-};
+} from "@/lib/location/locationLabels";
